requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11
django-cors-headers==2.0.2
djangorestframework==3.6.2
gunicorn==19.7.1
idna==2.5
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
psycopg2==2.7.3.2
PyMySQL==0.7.11
python-decouple==3.1
pytz==2017.3
requests==2.18.1
uritemplate==3.0.0
urllib3==1.21.1
whitenoise==3.3.1
